+++ CHANGES.rst
@@ -6,6 +6,8 @@ Unreleased (1.2.7) (XXXX-XX-XX)
 
 - Scenario table without table header and scrollable.
 
+- Add lookup filter for culvert and weir attributes.
+
 
 Release 1.2.25 (2015-2-19)
 ---------------------